devmode

+++ src/index.js
@@ -2,6 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import localForage from "localforage";
 
+import { log } from "./log";
+import { config } from "./config";
+
 import { initialState } from "./initialState";
 import { App } from "./App";
 import "./styles.scss";
@@ -14,11 +17,13 @@ localForage
       document.getElementById("app")
     );
   })
-  .catch(console.log);
+  .catch(log);
 
-window.clearSave = () => {
-  window.indexedDB.deleteDatabase("localforage");
-  location.reload();
-};
+if (config.isDev) {
+  window.clearSave = () => {
+    window.indexedDB.deleteDatabase("localforage");
+    window.location.reload();
+  };
+}
 
 module.hot.accept();
